% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materialstock_update.R
\name{materialstock_update_previewServer}
\alias{materialstock_update_previewServer}
\title{预览}
\usage{
materialstock_update_previewServer(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回测试数据集
}
\description{
预览
}
\examples{
materialstock_update_previewServer()
}
